Rename create_CC_mesh_from_contours to CCMesh.from_contours and make it a class method Remove vox2ras_tkr from fsaverage information Fix surface saving Fix --slice_selection type function Fix mapping from upright to orig space Change types of ndarrays from python to numpy Scalar types

+++ CorpusCallosum/segmentation/segmentation_postprocessing.py
@@ -500,9 +500,9 @@ def extract_largest_connected_component(
 
 
 def clean_cc_segmentation(
-    seg_arr: np.ndarray[Shape3d, np.dtype[int]],
+    seg_arr: np.ndarray[Shape3d, np.dtype[np.int_]],
     max_connection_distance: float = 3.0,
-) -> tuple[np.ndarray[Shape3d, np.dtype[int]], Mask3d]:
+) -> tuple[np.ndarray[Shape3d, np.dtype[np.int_]], Mask3d]:
     """Clean corpus callosum segmentation by removing non-connected components.
 
     Parameters
